Move WzCheckboxButton into separate files
