+++ src/notifications.cpp
@@ -352,33 +352,7 @@ void removeInGameNotificationForm(WZ_Queued_Notification* request);
 #endif
 #include <glm/gtx/transform.hpp>
 #include "lib/ivis_opengl/pieblitfunc.h"
-
-struct WzCheckboxButton : public W_BUTTON
-{
-public:
-	WzCheckboxButton() : W_BUTTON()
-	{
-		addOnClickHandler([](W_BUTTON& button) {
-			WzCheckboxButton& self = static_cast<WzCheckboxButton&>(button);
-			self.isChecked = !self.isChecked;
-		});
-	}
-
-	void display(int xOffset, int yOffset);
-
-	Vector2i calculateDesiredDimensions();
-
-	bool getIsChecked() const { return isChecked; }
-private:
-	int checkboxSize()
-	{
-		wzText.setText(pText.toUtf8(), FontID);
-		return wzText.lineSize() - 2;
-	}
-private:
-	WzText wzText;
-	bool isChecked = false;
-};
+#include "lib/widget/checkbox.h"
 
 static W_FORMINIT MakeNotificationFormInit()
 {
@@ -475,68 +449,6 @@ std::unique_ptr<WZ_Queued_Notification> popNextQueuedNotification()
 		++it;
 	}
 	return nullptr;
-}
-
-// MARK: - WzCheckboxButton
-
-Vector2i WzCheckboxButton::calculateDesiredDimensions()
-{
-	int cbSize = checkboxSize();
-	Vector2i checkboxPos{x(), y()};
-	int textLeftPos = checkboxPos.x + cbSize + 7;
-
-	// TODO: Incorporate padding?
-	return Vector2i(textLeftPos + wzText.width(), std::max(wzText.lineSize(), cbSize));
-}
-
-void WzCheckboxButton::display(int xOffset, int yOffset)
-{
-	wzText.setText(pText.toUtf8(), FontID);
-
-	int x0 = xOffset + x();
-	int y0 = yOffset + y();
-
-	bool down = (getState() & (WBUT_DOWN | WBUT_LOCK | WBUT_CLICKLOCK)) != 0;
-	bool grey = (getState() & WBUT_DISABLE) != 0;
-
-	// calculate checkbox dimensions
-	int cbSize = checkboxSize();
-	Vector2i checkboxOffset{0, (height() - cbSize) / 2}; // left-align, center vertically
-	Vector2i checkboxPos{x0 + checkboxOffset.x, y0 + checkboxOffset.y};
-
-	// draw checkbox border
-	PIELIGHT notifyBoxAddColor = WZCOL_NOTIFICATION_BOX;
-	notifyBoxAddColor.byte.a = uint8_t(float(notifyBoxAddColor.byte.a) * 0.7f);
-	pie_UniTransBoxFill(checkboxPos.x, checkboxPos.y, checkboxPos.x + cbSize, checkboxPos.y + cbSize, notifyBoxAddColor);
-	iV_Box2(checkboxPos.x, checkboxPos.y, checkboxPos.x + cbSize, checkboxPos.y + cbSize, WZCOL_TEXT_MEDIUM, WZCOL_TEXT_MEDIUM);
-
-	if (down || isChecked)
-	{
-		// draw checkbox "checked" inside
-		#define CB_INNER_INSET 2
-		PIELIGHT checkBoxInsideColor = WZCOL_TEXT_MEDIUM;
-		checkBoxInsideColor.byte.a = 200;
-		pie_UniTransBoxFill(checkboxPos.x + CB_INNER_INSET, checkboxPos.y + CB_INNER_INSET, checkboxPos.x + cbSize - (CB_INNER_INSET), checkboxPos.y + cbSize - (CB_INNER_INSET), checkBoxInsideColor);
-	}
-
-	if (grey)
-	{
-		// disabled, render something over it!
-		iV_TransBoxFill(x0, y0, x0 + width(), y0 + height());
-	}
-
-	// Display text to the right of the checkbox image
-	int textLeftPos = checkboxPos.x + cbSize + 7;
-	int fx = textLeftPos;
-	int fw = wzText.width();
-	int fy = yOffset + y() + (height() - wzText.lineSize()) / 2 - wzText.aboveBase();
-
-	if (style & WBUT_TXTCENTRE)							//check for centering, calculate offset.
-	{
-		fx = textLeftPos + ((width() - fw) / 2);
-	}
-
-	wzText.render(fx, fy, WZCOL_TEXT_MEDIUM);
 }
 
 // MARK: - W_NOTIFICATION
